Add files via upload
This is based on the SparkFun Serial Basic Breakout is an easy-to-use USB-to-Serial adapter based on the CH340E and takes advantage of the handy USB-C connector. With USB-C you can get up to three times the power delivery over the previous USB generation at 1.5A and and also solves the universally frustrating dilemma of plugging a USB cable in correctly, because it’s reversible! The Serial Basic works with 5V and 3.3V systems and possesses the capability to auto install on most operating systems without the need for additional drivers. The Serial Basic uses the CH340E IC to quickly and easily convert serial signals to USB.

The pinout of the Serial Basic mimics the common DTR/RXI/TXO/VCC/CTS/GND pinout found on hundreds of FTDI-to-USB derivatives. It can also be used as a general serial device for debugging (such as with a GPS module). There is a jumper on the rear of the board that controls the output voltage on the VCC pin. By default, the board outputs 3.3V and has 3.3V signals. Changing this jumper to 5V will cause the board to output 5V on the VCC pin with 5V signals.